auth-service/requirements.txt
dotenv
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pyjwt